Modularize booking page
