+++ src/styles/booking/form.css
@@ -1,0 +1,167 @@
+.booking-form,
+.booking-summary__card,
+.booking-summary__mini {
+  border: 1px solid color-mix(in srgb, currentColor 12%, transparent);
+  background: color-mix(in srgb, var(--surface) 94%, transparent);
+  box-shadow: 0 24px 58px -40px rgba(0,0,0,.55);
+}
+
+.booking-form {
+  display: grid;
+  gap: 1.25rem;
+  padding: clamp(1.45rem, 3vw, 2.25rem);
+  border-radius: 18px;
+  scroll-margin-top: calc(var(--nav-height) + 1.25rem);
+}
+
+.booking-fieldset {
+  border: 1px solid color-mix(in srgb, currentColor 9%, transparent);
+  border-radius: 16px;
+  padding: 1rem;
+  margin: 0;
+  display: grid;
+  gap: .9rem;
+  background: color-mix(in srgb, var(--page-bg) 20%, transparent);
+}
+
+.booking-fieldset legend,
+.booking-field > span {
+  font-family: var(--dp-font-sans);
+  color: var(--text-muted);
+  font-size: 11px;
+  font-weight: 800;
+  letter-spacing: .14em;
+  text-transform: uppercase;
+}
+
+.booking-fieldset legend {
+  padding: 0 .45rem;
+  margin-left: -.25rem;
+}
+
+.booking-choice-grid,
+.booking-payment-grid,
+.booking-transfer-tier-grid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: .75rem;
+}
+
+.booking-choice,
+.booking-payment {
+  display: grid;
+  gap: .35rem;
+  min-width: 0;
+  cursor: pointer;
+  border: 1px solid color-mix(in srgb, currentColor 13%, transparent);
+  border-radius: 12px;
+  padding: .95rem;
+  background: color-mix(in srgb, var(--surface) 70%, transparent);
+  transition: transform .25s ease, border-color .25s ease, background .25s ease, box-shadow .25s ease;
+}
+
+.booking-choice:hover,
+.booking-payment:hover,
+.booking-choice.is-selected,
+.booking-payment.is-selected {
+  transform: translateY(-2px);
+  border-color: color-mix(in srgb, var(--dp-accent) 58%, transparent);
+  background: color-mix(in srgb, var(--dp-accent) 9%, var(--surface));
+  box-shadow: 0 18px 36px -30px rgba(0,0,0,.5);
+}
+
+.booking-choice input,
+.booking-payment input {
+  position: absolute;
+  opacity: 0;
+  pointer-events: none;
+}
+
+.booking-choice span,
+.booking-payment span {
+  color: var(--ink);
+  font-family: var(--dp-font-display);
+  font-size: 1.02rem;
+  line-height: 1.25;
+  font-weight: 700;
+}
+
+.booking-choice small,
+.booking-payment small {
+  color: var(--text-muted);
+  font-family: var(--dp-font-display);
+  font-size: .86rem;
+  line-height: 1.45;
+}
+
+.booking-row {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 1rem;
+}
+
+.booking-row--thirds {
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+}
+
+.booking-field {
+  display: grid;
+  gap: .45rem;
+  min-width: 0;
+}
+
+.booking-field input,
+.booking-field select,
+.booking-field textarea {
+  width: 100%;
+  border: 1px solid var(--border);
+  border-radius: 10px;
+  background: var(--surface);
+  color: var(--text);
+  padding: .82rem 1rem;
+  font-family: var(--dp-font-display);
+  font-size: 15px;
+  outline: 0;
+  transition: border-color .25s ease, box-shadow .25s ease, background .25s ease;
+}
+
+.booking-field select {
+  appearance: auto;
+}
+
+.booking-field textarea {
+  resize: vertical;
+  min-height: 136px;
+  line-height: 1.55;
+}
+
+.booking-field input:focus,
+.booking-field select:focus,
+.booking-field textarea:focus {
+  border-color: var(--dp-accent);
+  box-shadow: 0 0 0 3px rgba(255,111,97,.15);
+}
+
+.booking-submit {
+  justify-self: start;
+  margin-top: .25rem;
+}
+
+.booking-status {
+  margin: 0;
+  padding: .75rem .9rem;
+  border-radius: 12px;
+  font-family: var(--dp-font-sans);
+  font-size: .9rem;
+}
+
+.booking-status--ok {
+  color: var(--dp-success);
+  background: var(--dp-success-bg);
+}
+
+.booking-status--err {
+  color: var(--dp-error);
+  background: var(--dp-error-bg);
+}
+
