+++ src/styles/booking/layout.css
@@ -1,0 +1,64 @@
+.booking-shell {
+  padding: clamp(3.5rem, 7vw, 6rem) clamp(1.25rem, 5vw, 6rem) clamp(4rem, 8vw, 7rem);
+}
+
+.booking-flow {
+  padding: clamp(3rem, 5vw, 4.5rem) clamp(1.25rem, 5vw, 6rem) 0;
+  background: var(--page-bg);
+  display: grid;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 1rem;
+  max-width: calc(1240px + (clamp(1.25rem, 5vw, 6rem) * 2));
+  margin: 0 auto;
+}
+
+.booking-flow article {
+  min-width: 0;
+  border: 1px solid color-mix(in srgb, currentColor 12%, transparent);
+  border-radius: 18px;
+  padding: clamp(1.15rem, 2.2vw, 1.6rem);
+  background:
+    linear-gradient(180deg, color-mix(in srgb, var(--surface) 95%, transparent), color-mix(in srgb, var(--surface) 82%, transparent));
+  box-shadow: 0 18px 42px -34px rgba(0,0,0,.42);
+}
+
+.booking-flow span {
+  display: inline-flex;
+  color: var(--dp-accent);
+  font-family: var(--dp-font-sans);
+  font-size: 11px;
+  font-weight: 800;
+  letter-spacing: .18em;
+  margin-bottom: .65rem;
+}
+
+.booking-flow h2 {
+  margin: 0 0 .35rem;
+  color: var(--ink);
+  font-family: var(--dp-font-display);
+  font-size: 1.35rem;
+  line-height: 1.2;
+}
+
+.booking-flow p {
+  margin: 0;
+  color: var(--text-muted);
+  font-size: .95rem;
+  line-height: 1.55;
+}
+
+.booking-intro {
+  max-width: 760px;
+  margin: 0 auto clamp(2rem, 4vw, 3rem);
+  text-align: center;
+}
+
+.booking-layout {
+  max-width: 1240px;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: minmax(0, 1.28fr) minmax(320px, .72fr);
+  gap: clamp(1.25rem, 3vw, 2rem);
+  align-items: start;
+}
+
