+++ src/styles/booking/responsive.css
@@ -1,0 +1,40 @@
+@media (max-width: 980px) {
+  .booking-flow {
+    grid-template-columns: 1fr;
+  }
+
+  .booking-layout {
+    grid-template-columns: minmax(0, 1fr) minmax(280px, .72fr);
+    gap: 1rem;
+  }
+}
+
+@media (max-width: 820px) {
+  .booking-layout {
+    grid-template-columns: 1fr;
+  }
+
+  .booking-summary {
+    position: static;
+  }
+}
+
+@media (max-width: 720px) {
+  .booking-hero__stats {
+    grid-template-columns: 1fr;
+    gap: .9rem;
+    width: min(320px, 100%);
+  }
+
+  .booking-choice-grid,
+  .booking-payment-grid,
+  .booking-transfer-tier-grid,
+  .booking-row,
+  .booking-row--thirds {
+    grid-template-columns: 1fr;
+  }
+
+  .booking-submit {
+    width: 100%;
+  }
+}
